Moving a few things from dashboard folder
