+++ public/app/features/dashboard/components/ShareModal/ShareModalCtrl.test.ts
@@ -1,5 +1,5 @@
 import config from 'app/core/config';
-import { LinkSrv } from 'app/features/dashboard/panellinks/link_srv';
+import { LinkSrv } from 'app/features/panel/panellinks/link_srv';
 import { ShareModalCtrl } from './ShareModalCtrl';
 
 describe('ShareModalCtrl', () => {
