+++ public/app/features/panel/all.ts
@@ -4,3 +4,5 @@ import './solo_panel_ctrl';
 import './query_ctrl';
 import './panel_editor_tab';
 import './query_editor_row';
+import './repeat_option';
+import './panellinks/module';
