+++ public/app/features/panel/panellinks/link_srv.ts
