+++ public/app/features/panel/panellinks/module.ts
