+++ public/app/features/panel/repeat_option.ts
